pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icedrive-authentication"
version = "0.3.5"
authors = [
  { name="Noelia Díaz Alejo Alejo", email="noelia.diaz14@alu.uclm.es" },
]
description = "Implementation of IceDrive Authentication service"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"
dependencies = [
    "zeroc-ice>=3.7",
]

[project.optional-dependencies]
tests = ["pytest"]
linters = ["black", "ruff", "pydocstyle"]

[project.urls]
"Homepage" = "https://github.com/NoeliaOwo/SDLab_Noelia"

[project.scripts]
icedrive-authentication-server = "icedrive_authentication.app:main"
icedrive-authentication-client = "client_testing.client:client"
icedrive-authentication-UIclient = "client_testing.UIclient:client"